requirements.txt
CHAID==5.4.2
graphviz==0.20.3
ipython==8.12.3
matplotlib==3.9.2
numpy==2.1.1
openml==0.14.2
pandas==2.2.2
Pillow==10.4.0
pydotplus==2.0.2
python_docx==1.1.2
pytorchcv==0.0.67
scikit_learn==1.4.2
scipy==1.14.1
seaborn==0.13.2
shap==0.45.1
sklearn2pmml==0.110.0
sklearn_pandas==2.2.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
tensorboard==2.17.1
